requirements.txt
fast-flights>=2.2
playwright>=1.40
requests>=2.31
python-dotenv>=1.0
